requirements.txt
torch
torchvision
numpy
pandas
opencv-python-headless
Pillow
scikit-learn
onnx
onnxruntime
matplotlib
seaborn
tqdm 